lint

+++ internal/database/collect/rand_test.go
@@ -1,3 +1,9 @@
+// Copyright 2023 The Accumulate Authors
+//
+// Use of this source code is governed by an MIT-style
+// license that can be found in the LICENSE file or at
+// https://opensource.org/licenses/MIT.
+
 package collect
 
 import (
@@ -83,5 +89,5 @@ func Test_LXRandom(t *testing.T) {
 	for i := 0; i < calls; i++ {
 		r.Hash()
 	}
-	fmt.Printf("Time %10s/s\n", humanize.Comma(int64(float64(calls)/float64(time.Since(start).Seconds()))))
+	fmt.Printf("Time %10s/s\n", humanize.Comma(int64(float64(calls)/time.Since(start).Seconds())))
 }
